requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
apispec==5.2.2
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
blinker==1.5
boto3==1.25.1
botocore==1.28.1
build==0.8.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
Deprecated==1.2.13
dnspython==2.2.1
durationpy==0.5
flasgger==0.9.5
Flask==2.2.2
flask-apispec==0.11.4
Flask-Cors==3.0.10
flask-http-middleware==0.1.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flask-serialize==2.1.1
Flask-SQLAlchemy==3.0.0
flask-swagger==0.2.14
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
hjson==3.1.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.1
jwt==1.3.1
kappa==0.6.0
kombu==5.2.4
mailjet-rest==1.3.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
mistune==2.0.4
packaging==21.3
pep517==0.13.0
permissive-dict==1.0.4
pip-tools==6.9.0
placebo==0.10.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
py==1.11.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.1.0
pymongo==4.3.2
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
raven==6.10.0
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
sentry-sdk==1.9.10
six==1.16.0
SQLAlchemy==1.4.41
swagger-ui==0.1.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
troposphere==4.1.0
unicorn==2.0.0
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
webargs==8.2.0
Werkzeug==2.2.2
wrapt==1.14.1
wsgi-request-logger==0.4.6
WTForms==3.0.1
zappa==0.54.2
zerobounce==0.1.1
